+++ tests/test_data/bigOvalTree.json
@@ -1,0 +1,76 @@
+{
+    "node_id": 1,
+    "type": "operator",
+    "value": "and",
+    "child": [
+        {
+            "node_id": 2,
+            "type": "value",
+            "value": "false",
+            "child": null
+        },
+        {
+            "node_id": 3,
+            "type": "operator",
+            "value": "xor",
+            "child": [
+                {
+                    "node_id": 4,
+                    "type": "value",
+                    "value": "true",
+                    "child": null
+                },
+                {
+                    "node_id": 5,
+                    "type": "operator",
+                    "value": "one",
+                    "child": [
+                        {
+                            "node_id": 6,
+                            "type": "value",
+                            "value": "noteval",
+                            "child": null
+                        },
+                        {
+                            "node_id": 7,
+                            "type": "value",
+                            "value": "true",
+                            "child": null
+                        },
+                        {
+                            "node_id": 8,
+                            "type": "value",
+                            "value": "notappl",
+                            "child": null
+                        }
+                    ]
+                },
+                {
+                    "node_id": 9,
+                    "type": "value",
+                    "value": "error",
+                    "child": null
+                }
+            ]
+        },
+        {
+            "node_id": 10,
+            "type": "operator",
+            "value": "or",
+            "child": [
+                {
+                    "node_id": 11,
+                    "type": "value",
+                    "value": "unknown",
+                    "child": null
+                },
+                {
+                    "node_id": 12,
+                    "type": "value",
+                    "value": "true",
+                    "child": null
+                }
+            ]
+        }
+    ]
+}
